New sliding pane layout combining conversation and contact list
ContactsListActivity remains as picker activity.
