+++ app/src/main/AndroidManifest.xml
@@ -103,7 +103,8 @@
         <!-- Activities -->
         <activity android:name=".ui.ConversationList"
             android:launchMode="singleTop"
-            android:theme="@style/Theme.Light">
+            android:theme="@style/Theme.Light"
+            android:configChanges="orientation|screenSize">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
