+++ app/src/main/java/org/kontalk/ui/ComposeMessage.java
@@ -376,7 +376,6 @@ public class ComposeMessage extends ActionBarActivity {
         // TODO one day it will be like this
         // Intent i = new Intent(Intent.ACTION_PICK, Users.CONTENT_URI);
         Intent i = new Intent(this, ContactsListActivity.class);
-        i.putExtra("picker", true);
         startActivityForResult(i, REQUEST_CONTACT_PICKER);
     }
